Move eicon search eicon load back to webview for performance, submit metadata to host
